Fix point balance for datatype and return categories with product in resource

+++ database/migrations/2025_06_19_183817_add_points_balance_in_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('points_balance')->unsigned()->default(0);
+            $table->bigInteger('points_balance')->unsigned()->default(0);
         });
     }
